Fix issue with logic to dedup randomly-chosen images
I neglected to put a 'continue' after the previous logic to add
one to the index offset, so the index it was checking was always
0. This led to an infinite loop, since I didn't have any guardrails
on the while.
This change adds guardrails, neatens the logic a little with a
modulus operator, and correctly iterates over the list of all
images until a new image is found.